Extract highlights for specific O'Reilly book
O'Reilly provide a CSV file containing all the highlights you've made to
any of the books that you read on their online learning site. [1]

This script is a quick hack to generate a Markdown file containing the
highlights from a specific book, that I can then import into Notion.

Run it like this:

    $ ./oreilly-book "Book Title" < safari-annotations-export.csv > book.md

The "Book Title" string must match the first column in the CSV file for
the book you're extracting highlights for.

[1] https://learning.oreilly.com